requirements.txt
arrow==1.0.3
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
drfpasswordless==1.5.7
furl==2.1.0
idna==2.10
inflection==0.5.1
jsonschema==3.2.0
orderedmultidict==1.0.1
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
psycopg2-binary==2.8.6
Pillow==8.1.2
